Fixed some bit rot issues

+++ LocalTimeServer.cpp
@@ -30,7 +30,8 @@ DAMAGE.
 
 #include "mil/LocalTimeServer.h"
 
-#include <ESP8266HTTPClient.h>
+#include <WiFi.h>
+#include <HTTPClient.h>
 #include <JsonStreamingParser.h>
 #include <time.h>
 
@@ -97,7 +98,7 @@ bool LocalTimeServer::update()
 	}
 
 	if (httpCode > 0) {
-		mil::cout << F("    got response: ") << httpCode << F("\n");
+		mil::cout << F("    got response: ") << int32_t(httpCode) << F("\n");
 
 		if(httpCode == HTTP_CODE_OK) {
 			String payload = http.getString();
@@ -112,7 +113,7 @@ bool LocalTimeServer::update()
 			_currentTime = listener.localEpoch();
 		}
 	} else {
-		mil::cout << F("[HTTP] GET... failed, error: ") << http.errorToString(httpCode) << F("(") << httpCode << F(")\n");
+		mil::cout << F("[HTTP] GET... failed, error: ") << http.errorToString(httpCode) << F("(") << int32_t(httpCode) << F(")\n");
 		failed = true;
 	}
